+++ test/zuul-cli.js
@@ -1,16 +1,14 @@
 import test from 'ava';
 import execa from 'execa';
 
-test.skip('zuul', t => {
-  process.chdir('../');
+process.chdir('../');
 
-  return execa.shell('zuul --ui tape --phantom -- ./test/fixtures/assert/concat.js', {preferLocal: true})
-    .then(result => {
-      t.regex(result.stdout, /TAP version/);
-      t.regex(result.stdout, /# tests 3/);
-      t.regex(result.stdout, /# pass {2}3/);
-      t.regex(result.stdout, /## array slice/);
-      t.regex(result.stdout, /## array slice - without asserts/);
-      t.regex(result.stdout, /\*Fastest is __for loop__\*/);
-    });
+test.skip('zuul', async t => {
+  const result = await execa.shell('zuul --ui tape --phantom -- ./test/fixtures/assert/concat.js', {preferLocal: true});
+  t.regex(result.stdout, /TAP version/);
+  t.regex(result.stdout, /# tests 3/);
+  t.regex(result.stdout, /# pass {2}3/);
+  t.regex(result.stdout, /## array slice/);
+  t.regex(result.stdout, /## array slice - without asserts/);
+  t.regex(result.stdout, /\*Fastest is __for loop__\*/);
 });